ai-consumer/pyproject.toml
[project]
name = "ai-consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.2.15",
    "openai>=2.1.0",
    "pydantic-settings>=2.11.0",
]

[dependency-groups]
dev = [
    "logger",
]

[tool.uv.sources]
logger = { path = "../common/logger" }
